Основной сервер: Пройденных тестов: 229 Упавших тестов: 3

+++ models/src/main/java/ru/practicum/eventDto/EventFullDto.java
@@ -33,6 +33,6 @@ public record EventFullDto(
         String state,
         @NotNull
         String title,
-        Long views
+        Integer views
 ) {
 }
